Split MCP connection modules
Split the former MCP connection manager module into focused modules for manager orchestration, RMCP client startup, elicitation handling, Codex Apps support, tool helpers, and runtime support.

Preserve latest main's PermissionProfile-based MCP elicitation approval behavior while applying the module split.
