+++ codex-rs/codex-mcp/src/mcp_connection_manager_tests.rs
@@ -1,12 +1,36 @@
 use super::*;
+use crate::apps::CODEX_APPS_TOOLS_CACHE_SCHEMA_VERSION;
+use crate::apps::CodexAppsToolsCacheContext;
+use crate::apps::load_startup_cached_codex_apps_tools_snapshot;
+use crate::apps::read_cached_codex_apps_tools;
+use crate::apps::write_cached_codex_apps_tools;
+use crate::client::AsyncManagedClient;
+use crate::client::ManagedClient;
+use crate::client::StartupOutcomeError;
+use crate::client::elicitation_capability_for_server;
+use crate::declared_openai_file_input_param_names;
+use crate::elicitation::ElicitationRequestManager;
+use crate::elicitation::elicitation_is_rejected_by_policy;
+use crate::tools::ToolFilter;
+use crate::tools::ToolInfo;
+use crate::tools::filter_tools;
+use crate::tools::qualify_tools;
+use crate::tools::tool_with_model_visible_input_schema;
+use codex_config::Constrained;
 use codex_protocol::ToolName;
 use codex_protocol::models::PermissionProfile;
 use codex_protocol::protocol::GranularApprovalConfig;
 use codex_protocol::protocol::McpAuthStatus;
+use futures::FutureExt;
 use pretty_assertions::assert_eq;
+use rmcp::model::CreateElicitationRequestParams;
+use rmcp::model::ElicitationAction;
+use rmcp::model::ElicitationCapability;
+use rmcp::model::FormElicitationCapability;
 use rmcp::model::JsonObject;
 use rmcp::model::Meta;
 use rmcp::model::NumberOrString;
+use rmcp::model::Tool;
 use std::collections::HashSet;
 use std::sync::Arc;
 use tempfile::tempdir;
